Update README.rst
It was quite confusing as to why there's the possibility of events being still populated after the given loop.  I get that in a more complex example, there may still be events left over, but I think the example was misleading so am proposing just asserting that the events are empty which might serve better to explain the workings.